Commit changes

+++ lesson_seven/src/energy/form.rs
@@ -1,7 +1,4 @@
-use crate::liquid;
-use crate::energy;
-
 pub struct Temperature {
-    pub value: f32,
+    pub value: i16,
     pub unit: String
 }
